chore: harden sdk packaging surface
